Allow BLEMRI to handle more than GDB packets
Previously the BLEMRI firmware running on the nRF51 expected GDB remote
packets to be sent so it would buffer data until it was full or the end
of a packet was received. This commit removes that limitation. Now it
buffers up the data until the buffer is full or no UART data has been
sent from the LPC1768 for 1.5 serial frames. This fixes an issue where
it failed to send the GDB remote protocol ack ('+') for the continue
command which resulted in GDB timing out and sending the continue
command again. This would cause MRI to break into the running LPC1768.