Update LX16A sample
I updated the LX16A test sample to change the active servo Id to match
the returned Id when running option "1. Discover Servo ID when
only 1 servo is attached." This makes it easier to communicate with a
servo after initially connecting it by just running option 1 and then
any other option will be setup to use the discovered servo.