Share same sdk.patch across nRF51 based projects
Previously each nRF51 based project had its own sdk.patch file to use
in the installation of the Nordic nRF5 SDK. This change places in a
common folder for all of those projects so that they all apply the
same changes, no matter which project is used for the SDK install.

I also updated sdk.patch to patch Makefile.common to dump the
disassembly of the final binaries.