Add deadman switch to BleJoystick prototype
I moved the joystick prototype over to a breadboard which uses the
Adafruit DragonTail adapter to more easily access more pins on the
micro:bit. Once I had access to more pins, I updated the BleJoystick
firmware and the macosTest code to support the deadman switch in
addition to presses of the joystick itself.