Close relay when remote deadman switch pressed
I connected the relay control input to P8 (P0.18) of the micro:bit and
used it to close the motor power relay when the deadman switch was
depressed on the remote BLE unit. The BJT controlled by this digital
output requires a base current of 5mA so P8 was configured for high
current drive. I also needed to modify the Sparkfun Beefcake Relay
Control Board to replace the 1k base resistor with a 680 ohm unit so
that enough CE current was allowed to flow through the relay when using
a 3.3V digital output. Without this change, the relay wouldn't fully
close when the BJT was turned on.

I also commented out most of the LCD display test code as I wanted to
not wait too long before I could test my relay setup and code.