Update blemri firmware
I used the new C structure initializer format with named struct fields
for setting up the UART connection parameters to make it easier to
read. I had used this format in some of my newer nRF51 code so I
backported it to this project too.

I also fixed a bug in its Makefile to pull the programming tools from
the correct folder. I had already fixed the other nrfjprog
instantiations but missed the one for the "flash" rule.