Initial commit of fast ST7789 display lib
This is the library that I initially pulled down from @cbm80amiga 's
GitHub repository in an attempt to speed up the 1.54" TFT display that
I am currently using with the power distribution board. The GitHub link
is:
  https://github.com/cbm80amiga/Arduino_ST7789_Fast

Later commits will come to make it build for the nRF51
microcontroller.