Initial commit of RREFont library
This is the library that I initially pulled down from @cbm80amiga 's
GitHub repository in an attempt to speed up the 1.54" TFT display that
I am currently using with the power distribution board. The GitHub link
is:
  https://github.com/cbm80amiga/RREFont

Later commits will come to make it build for the nRF51
microcontroller.