RREFont updates
Most of the changes were to just get it to build for nRF51. I did
switch from a callback for the fillRect function and used an interface
instead so that I didn't need a global function. I added this interface
to the Arduino_SST7789 class. I also used const char* pointers instead
of char* pointer where possible.