Faster detection of BLE remote control disconnect
The BLE central device doesn't consider a BLE peripheral to be
disconnected for a few seconds after it stops seeing radio traffic from
the device. This improves robustness to short bursts of radio
interference but in our case, we want the motors to be disabled as soon
as the radio connection is broken. Otherwise it might miss us releasing
the deadman switch if there was radio interference at the same time.

I refactored the code to have the LCD update timer code check to see if
the BLE stack has received new packets since its previous iteration. If
it doesn't see any BLE traffic for 2 iterations of the LCD timer then
it considers the remote to have been disconnected and disables the
motor power.

The LCD update timer now works on local copies of the PDB and remote
control state so that it sees a consistent view of the state throughout
a single iteration of the timer. This allowed me to remove the PdbState
copy that was taken in Screen::update for similar reasons.