Blink battery voltages when low
I refactored the code to move the existing blink code out into its own
Blinker helper class and then use it to make the battery voltages blink
as well when they aren't green.

I now calculate the elapsed time since the beginning of this blink
cycle into Screen::update() where it is done once instead of for each
indicator that is in the process of blinking.