Archive of PDB port to nRF52 based Feather
I didn't bother to get the battery voltage level reading code to work
with the differences in the ADC peripheral on the nRF52 but this is the
only thing I know that still needed to be addressed in this port. I
don't actually plan to switch to a nRF52 based board at this time so I
am just archiving away what I have so far incase I want to revisit in
the future.