Sample rate now in config.ini
Can now set the sample rate in config.ini rather than be hardcoded in
the Java and C++ source code. 200Hz seems to work ok but 400Hz turns
out to be too fast for the UART to keep up when sending the full larger
packets. Would work on the robot itself though since it wouldn't need
to send the data out over the UART.

NOTE: I also updated the compass.gyro.scale in config.ini to just be
      the 16.0 value from the datasheet instead of my badly measured
      values.