Improve blemri throughput back to PC
The code now queue up as many of the chunks directly to the BLE stack
until it gets a BLE_ERROR_NO_TX_PACKETS error. Once that error is
received, the packets are placed in the chunk queue which drains as the
BLE stack completes the sending of previous chunks. Once this queue is
drained, the code will go back to submitting directly to the BLE stack
again.

I can now successfully use the DebugSerial class to send debug spew
back to GDB and GDB feels much more responsive after I made this
change since it lets the BLE stack know that it can transmit more
than 1 chunk per connection interval.