Move include/ and libs/ back into software/
That way they are all covered by the same LICENSE.TXT in software/. I
also placed a copy of LICENSE.TXT in the processing/ folder as well.

I also moved EncoderTest into libs/LX16A since it is a test for that
library and I probably won't revisit it a lot in the future.